backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
pandas==2.0.3
numpy==1.24.4
scikit-learn==1.3.2
joblib==1.4.2
